New CLI for img-to-shp.py issue-85
